[refactor] Hook 중복 제거
- Modal 창 닫힘/열림 관련 Hook으로 독립되게 수정함
- Input Text 상태 관리는 기존에 있던 Hook을 재사용해서 사용할 수 있도록 수정
