+++ src/hooks/useInputText.ts
@@ -1,13 +1,14 @@
-import { useState, ChangeEvent } from 'react';
+import React, { useState, ChangeEvent } from 'react';
 
+type InputType = HTMLInputElement | HTMLTextAreaElement;
 export interface InputTextHookType {
   inputText: string;
-  onInputChange: (e: ChangeEvent<HTMLInputElement>) => void;
+  onInputChange: (e: ChangeEvent<InputType>) => void;
 }
 
 function useInputText(): InputTextHookType {
   const [inputText, setInputText] = useState<string>('');
-  const onInputChange = (e: ChangeEvent<HTMLInputElement>) => {
+  const onInputChange = (e: React.ChangeEvent<InputType>) => {
     const newInputText = e.target.value;
     setInputText(newInputText);
   };
